feat: show all the changelogs on changelog section
We have already been through the times when we push an update with lots of new features and fixes but due to some bugs, we immediately push a hotfix to resolve the issue. However, because of that hotfix, the changelog gets updated which just mentions about the hotfix and the users who didn't update the app before the hotfix won't be able to see all the new changes within the app while updating.

This PR fixes that problem by merging the changelogs.

For example: If I am on `v1.3.0`, then, I will be able to see the changelogs from `v1.3.1` upto the latest version.