fix: default to non-multithreaded until ui toggle support
Patching large applications (e.g. TikTok) on Android with multithreading simply doesn't work on multicore devices where the amount of memory given to the manager (regardless of the system RAM) is not sufficient.

It's more important that patching succeeds for all available use cases than for only some patching to succeed with faster performance. A UI option can be added later, but the multithreading option should be opt-out by default rather than opt-in.

Closes ReVanced/revanced-documentation#35
Closes ReVanced/revanced-manager#1454
Closes ReVanced/revanced-manager#1571
Closes ReVanced/revanced-manager#1595
Closes ReVanced/revanced-manager#1659
Closes ReVanced/revanced-manager#1661
Closes ReVanced/revanced-manager#1684
Closes ReVanced/revanced-manager#1759
Closes ReVanced/revanced-manager#1802
Closes ReVanced/revanced-manager#1817
Closes ReVanced/revanced-patches#2885
Closes ReVanced/revanced-manager#592
Closes ReVanced/revanced-patcher#193
Closes ReVanced/revanced-patches#1533
Closes ReVanced/revanced-patches#1608
Closes ReVanced/revanced-patches#1613
Closes ReVanced/revanced-patches#1630
Closes ReVanced/revanced-patches#190
Closes ReVanced/revanced-patches#2511
Closes ReVanced/revanced-patches#525