fix: Resolve EACCESS error in special cases
Sometimes, when there is an attempt to patch an App, that APK is only marked as `r-x` and not `rwx`.
This can cause the patch to fail, as it cannot read the file after copying it to `patcher/tmp-.../in.apk`
This PR Fixes that issue (as far as i can tell; tested with brand new install of revanced manager, with and without patch. to Verify)