chore: Add ads.fund verification file [skip ci]
This file is intended for the ReVanced/.github repository; however, ads.fund is responsible for configuring the verification for this repository. I have contacted them to set up the .github repository, but they are currently offline. Therefore, this file is committed here temporarily due to a time constraint, as the verification must be completed before the end of this month. Once the verification process is finished and the changes on the ads.fund side are implemented, this file can be removed.